Another take on redoing the streams design/implementation
