+++ src/actor/mod.rs
@@ -12,11 +12,8 @@ use crate::mailbox::*;
 use crate::timer::*;
 
 pub use self::actor_ref::{
-    Actor, ActorContext, ActorRef, ActorSpawnContext, FailureAction, FailureError, FailureReason, Signal,
-    Spawnable,
-    StopReason,
-    SystemActorRef,
-    Watchable
+    Actor, ActorContext, ActorRef, ActorSpawnContext, FailureAction, FailureError, FailureReason,
+    Signal, Spawnable, StopReason, SystemActorRef, Watchable,
 };
 pub use self::probe::{Probe, SpawnProbe};
 pub use self::system::{ActiveActorSystem, ActorSystem, ActorSystemContext};
