+++ src/stream/sink/ignore.rs
@@ -1,95 +1,47 @@
-use crate::dispatcher::{BoxedFn1In0Out, Trampoline};
-use crate::stream::oxidized::*;
-use crate::stream::*;
+use crate::stream::{Action, Logic, StreamContext};
 use std::marker::PhantomData;
-use std::panic::{catch_unwind, UnwindSafe};
 
-/// `Ignore` is a `Consumer` that requests elements
-/// one at a time and drops them as it receives them.
-///
-/// It can be connected to producers to constantly
-/// pull values from upstream.
-pub struct Ignore<A, Up: Producer<A>>
+struct Ignore<A>
 where
     A: 'static + Send,
 {
-    on_termination: Option<Box<dyn BoxedFn1In0Out<Terminated> + 'static + Send + UnwindSafe>>,
-    upstream: Up,
+    pulled: bool,
     phantom: PhantomData<A>,
 }
 
-impl<A, Up: Producer<A>> Ignore<A, Up>
+impl<A> Logic<A, (), ()> for Ignore<A>
 where
     A: 'static + Send,
 {
-    pub fn watch_termination<F: FnOnce(Terminated)>(mut self, f: F) -> Self
-    where
-        F: 'static + Send + UnwindSafe,
-    {
-        self.on_termination = Some(Box::new(f));
-        self
+    fn name(&self) -> &'static str {
+        "Ignore"
     }
-}
 
-impl<A, Up: Producer<A>> Ignore<A, Up>
-where
-    A: 'static + Send,
-{
-    pub fn new() -> impl FnOnce(Up) -> Self {
-        move |upstream| Self {
-            on_termination: None,
-            upstream,
-            phantom: PhantomData,
+    fn pulled(&mut self, ctx: &mut StreamContext<A, (), ()>) -> Option<Action<(), ()>> {
+        self.pulled = true;
+
+        None
+    }
+
+    fn pushed(&mut self, el: A, ctx: &mut StreamContext<A, (), ()>) -> Option<Action<(), ()>> {
+        Some(Action::Pull)
+    }
+
+    fn started(&mut self, ctx: &mut StreamContext<A, (), ()>) -> Option<Action<(), ()>> {
+        Some(Action::Pull)
+    }
+
+    fn stopped(&mut self, ctx: &mut StreamContext<A, (), ()>) -> Option<Action<(), ()>> {
+        if self.pulled {
+            self.pulled = false;
+
+            Some(Action::Push(()))
+        } else {
+            None
         }
     }
-}
 
-impl<A, Up: Producer<A>> Consumer<A> for Ignore<A, Up>
-where
-    A: 'static + Send,
-{
-    fn started<Produce: Producer<A>>(self, producer: Produce, _: &StreamContext) -> Trampoline {
-        producer.pull(self)
-    }
-
-    fn produced<Produce: Producer<A>>(self, producer: Produce, _: A) -> Trampoline {
-        Trampoline::bounce(move || producer.pull(self))
-    }
-
-    fn completed(self) -> Trampoline {
-        if let Some(f) = self.on_termination {
-            if let Err(_e) = catch_unwind(move || f.apply(Terminated::Completed)) {
-                // @TODO we should have some mechanism for this
-                debug!("user-supplied watch_termination function panicked");
-            }
-        }
-
-        Trampoline::done()
-    }
-
-    fn failed(self, e: Error) -> Trampoline {
-        if let Some(f) = self.on_termination {
-            if let Err(_e) = catch_unwind(move || f.apply(Terminated::Failed(e))) {
-                // @TODO we should have some mechanism for this
-                debug!("user-supplied watch_termination function panicked");
-            }
-        }
-
-        Trampoline::done()
-    }
-}
-
-impl<A, Up: Producer<A>> Sink<A> for Ignore<A, Up>
-where
-    A: 'static + Send,
-{
-    fn start(self, stream_context: &StreamContext) -> Trampoline {
-        let sink = Ignore {
-            on_termination: self.on_termination,
-            upstream: Disconnected,
-            phantom: PhantomData,
-        };
-
-        self.upstream.attach(sink, &stream_context)
+    fn forwarded(&mut self, msg: (), ctx: &mut StreamContext<A, (), ()>) -> Option<Action<(), ()>> {
+        None
     }
 }
