+++ src/stream/sink/mod.rs
@@ -1,69 +1,63 @@
-pub mod cancel;
-pub mod detached_logic;
+use crate::stream::internal::{
+    InternalStreamCtl, LogicContainer, LogicContainerFacade, ProtectedStreamCtl,
+};
+use crate::stream::Logic;
+use std::marker::PhantomData;
+
+pub mod collect;
 pub mod first;
 pub mod for_each;
 pub mod ignore;
 pub mod last;
-pub mod tell;
 
-pub use cancel::Cancel;
-pub use first::First;
-pub use for_each::ForEach;
-pub use ignore::Ignore;
-pub use last::Last;
-pub use tell::Tell;
+/// A `Sink` is a stage that accepts a single output, and outputs a
+/// terminal value.
+///
+/// The logic supplied for a `Sink` will be pulled immediately when
+/// the stream is spawned. Once the stream has finished, the logic's
+/// stop handler will be invoked, and a conforming implementation must
+/// at some point push out a single value.
+pub struct Sink<A, Out>
+where
+    Out: 'static + Send,
+{
+    pub(in crate::stream) logic: Box<LogicContainerFacade<A, Out, InternalStreamCtl<Out>> + Send>,
+    phantom: PhantomData<Out>,
+}
 
-use crate::actor::ActorRef;
-use crate::stream::sink::detached_logic::DetachedLogicSink;
-use crate::stream::Stage;
-use tell::{TellCommand, TellEvent};
-
-/// Contains methods to construct commonly used `Sink` types.
-pub struct Sinks;
-
-impl Sinks {
-    pub fn cancel<A, Up: Stage<A>>() -> impl FnOnce(Up) -> Cancel<A, Up>
+impl<In, Out> Sink<In, Out>
+where
+    In: 'static + Send,
+    Out: 'static + Send,
+{
+    pub fn new<Msg, L: Logic<In, Out, Msg>>(logic: L) -> Self
     where
-        A: 'static + Send,
+        Msg: 'static + Send,
+        L: 'static + Send,
     {
-        Cancel::new()
+        Self {
+            logic: Box::new(LogicContainer {
+                logic,
+                phantom: PhantomData,
+            }),
+            phantom: PhantomData,
+        }
+    }
+}
+
+// @TODO redo this
+impl<A> Sink<A, ()>
+where
+    A: 'static + Send,
+{
+    pub fn first() -> Sink<A, Option<A>> {
+        Sink::new(first::First::new())
     }
 
-    pub fn first<A, Up: Stage<A>>() -> impl FnOnce(Up) -> First<A, Up>
+    pub fn for_each<F: FnMut(A) -> ()>(for_each_fn: F) -> Sink<A, ()>
     where
-        A: 'static + Send,
-    {
-        First::new()
-    }
-
-    pub fn for_each<A, F: FnMut(A), Up: Stage<A>>(func: F) -> impl FnOnce(Up) -> ForEach<A, F, Up>
-    where
-        A: 'static + Send,
         F: 'static + Send,
     {
-        ForEach::new(func)
-    }
-
-    pub fn ignore<A, Up: Stage<A>>() -> impl FnOnce(Up) -> Ignore<A, Up>
-    where
-        A: 'static + Send,
-    {
-        Ignore::new()
-    }
-
-    pub fn last<A, Up: Stage<A>>() -> impl FnOnce(Up) -> Last<A, Up>
-    where
-        A: 'static + Send,
-    {
-        Last::new()
-    }
-
-    pub fn tell<A, Up: Stage<A>>(
-        actor_ref: ActorRef<TellEvent<A>>,
-    ) -> impl FnOnce(Up) -> DetachedLogicSink<A, TellCommand, Tell<A>, Up>
-    where
-        A: 'static + Send,
-    {
-        DetachedLogicSink::new(Tell::new(actor_ref))
+        Sink::new(for_each::ForEach::new(for_each_fn))
     }
 }
